+++ example/lib/main.dart
@@ -30,8 +30,14 @@ class _MyAppState extends State<MyApp> {
   @override
   void initState() {
     _alice = Alice(
-        showNotification: true, showInspectorOnShake: true, darkTheme: false);
-    _dio = Dio(BaseOptions(followRedirects: false));
+      showNotification: true,
+      showInspectorOnShake: true,
+      darkTheme: false,
+      maxCallsCount: 1000
+    );
+    _dio = Dio(BaseOptions(
+      followRedirects: false,
+    ));
     _dio.interceptors.add(_alice.getDioInterceptor());
     _httpClient = HttpClient();
     _chopper = ChopperClient(